refactor(nightscout): update component names and imports
