+++ src/lib/components/v2/widgets/Nightscout/NightscoutGraph.svelte
